+++ geary-conventions/src/main/kotlin/geary.kotlin-conventions.gradle.kts
@@ -16,6 +16,8 @@ object GearyDeps {
     val bimap = "com.uchuhimo:kotlinx-bimap:1.2"
     val bitvector = "net.onedaybeard.bitvector:bitvector-jvm:0.1.4"
     val fastutil = "it.unimi.dsi:fastutil:8.5.4"
+    val reflections = "org.reflections:reflections:0.9.12"
+    val plugman = "com.rylinaux:PlugMan:2.2.5"
 }
 
 dependencies {
