+++ geary-core/src/main/kotlin/com/mineinabyss/geary/ecs/events/RequestCheck.kt
@@ -1,6 +1,5 @@
 package com.mineinabyss.geary.ecs.events
 
-//TODO cancelable events
 /**
  * A component attached to an event to request handler to check whether it should be successful.
  */
