--- geary-prefabs/src/main/kotlin/com/mineinabyss/geary/prefabs/GearyPrefabs.kt
@@ -1,4 +1,0 @@
-package com.mineinabyss.geary.prefabs
-
-public class GearyPrefabs {
-}
