--- geary-prefabs/src/main/kotlin/com/mineinabyss/geary/prefabs/PersistingPrefabs.kt
@@ -1,5 +1,0 @@
-package com.mineinabyss.geary.prefabs
-
-public data class PersistingPrefabs(
-    val prefabs: Set<PrefabKey>
-)
