+++ geary-prefabs/src/main/kotlin/com/mineinabyss/geary/prefabs/PrefabKey.kt
@@ -1,6 +1,7 @@
 package com.mineinabyss.geary.prefabs
 
 import com.mineinabyss.geary.ecs.api.entities.GearyEntity
+import com.mineinabyss.geary.prefabs.serializers.PrefabKeySerializer
 import kotlinx.serialization.Serializable
 
 /**
