+++ geary-prefabs/src/main/kotlin/com/mineinabyss/geary/prefabs/configuration/components/Prefab.kt
@@ -3,7 +3,7 @@ package com.mineinabyss.geary.prefabs.configuration.components
 import java.io.File
 
 /**
- * A component applied to prefabs loaded from a file that allows them to be reused.
+ * A component applied to prefabs loaded from a file that allows them to be reread.
  */
 public class Prefab(
     public val file: File
