+++ geary-prefabs/src/main/kotlin/com/mineinabyss/geary/prefabs/configuration/components/RelationOnPrefab.kt
@@ -5,6 +5,9 @@ import kotlinx.serialization.Polymorphic
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
+/**
+ * A component that will add a relation to this entity with a [key], [value] pair.
+ */
 @Serializable
 @SerialName("geary:relation")
 public data class RelationOnPrefab(
