+++ geary-prefabs/src/main/kotlin/com/mineinabyss/geary/prefabs/configuration/systems/ParseRelationOnPrefab.kt
@@ -1,11 +1,10 @@
-package com.mineinabyss.geary.prefabs
+package com.mineinabyss.geary.prefabs.configuration.systems
 
 import com.mineinabyss.geary.ecs.accessors.TargetScope
 import com.mineinabyss.geary.ecs.accessors.building.get
 import com.mineinabyss.geary.ecs.api.autoscan.AutoScan
 import com.mineinabyss.geary.ecs.api.autoscan.Handler
 import com.mineinabyss.geary.ecs.api.systems.GearyListener
-import com.mineinabyss.geary.ecs.components.RelationComponent
 import com.mineinabyss.geary.ecs.serialization.parseEntity
 import com.mineinabyss.geary.prefabs.configuration.components.RelationOnPrefab
 
@@ -27,17 +26,3 @@ public class ParseRelationOnPrefab : GearyListener() {
     }
 }
 
-@AutoScan
-public class ParseRelationComponent : GearyListener() {
-    private val TargetScope.relation by get<RelationComponent>()
-
-    init {
-        allAdded()
-    }
-
-    @Handler
-    private fun TargetScope.convertToRelation() {
-        entity.setRelation(relation.key, relation.value)
-        entity.remove<RelationComponent>()
-    }
-}
